+++ src/Client/AndroidUI.h
@@ -10,6 +10,8 @@
 #include <Geode/modify/CCScene.hpp>
 #include <Geode/modify/CCNode.hpp>
 
+#include "../AttemptAtReversingDialogObject.h"
+
 using namespace geode::prelude;
 
 class AndroidUI : public cocos2d::CCLayerColor {
@@ -21,6 +23,8 @@ public:
 
     static inline int lastTab = 0;
     static inline int selectedTab = 0;
+
+    int secret = 0;
     
     void goToPage(int p, bool transition = false)
     {
@@ -92,14 +96,101 @@ public:
 
         if (btn->getTag() == 5)
         {
-            FLAlertLayer::create("Coming soon", "Replay is not ready...", "Ok")->show();
-            //auto dO = DialogObject::create("Shopkeeper", "Hewwo :3", 1, 1.0f, true, {255, 255, 255});
-            //auto dialog = DialogObject::create("Scratch", "<cr><s005>RobTop</s></c>", 13, 1.0f, true, _ccColor3B({255, 255, 255}));
+            CCArray* arr = CCArray::create();
 
-            // auto layer = DialogLayer::create(dO, 0);
-            //layer->setZOrder(999999);
-            //CCScene::get()->addChild(layer);
-            //layer->animateIn(DialogAnimationType::FromCenter);
+            AttemptAtReversingDialogObject* Object = nullptr;
+            std::stringstream ss; // why do i have to declare this outside? seems stupid
+
+            switch (secret)
+            {
+            case 0:
+                Object = AttemptAtReversingDialogObject::create("The Shopkeeper", "A replay bot is planned, <d050>But I still need more time to develop it.", 5, 1, true, { 255,255,255 });
+                arr->addObject(Object);   
+
+                break;
+
+            case 1:
+                Object = AttemptAtReversingDialogObject::create("The Shopkeeper", "It still isn't finished, <d050>wait a bit.", 6, 1, true, { 255,255,255 });
+                arr->addObject(Object);   
+
+                break;
+
+            case 2:
+                Object = AttemptAtReversingDialogObject::create("The Shopkeeper", "I just said, <d050>it isn't done yet.", 31, 1, true, { 255,255,255 });
+                arr->addObject(Object);   
+
+                break;
+
+            case 3:
+                Object = AttemptAtReversingDialogObject::create("The Shopkeeper", "I'm warning you!", 30, 1, true, { 255,255,255 });
+                arr->addObject(Object);   
+
+                break;
+
+            case 4:
+                FMODAudioEngine::sharedEngine()->playMusic("dangerLoop.mp3", true, 0, 0);
+
+                Object = AttemptAtReversingDialogObject::create("The Shopkeeper", "STOP! <d050>ASKING!", 34, 1, true, { 255,255,255 });
+                arr->addObject(Object);   
+
+                Object = AttemptAtReversingDialogObject::create("The Shopkeeper", "Oh no! <d100>Just<d010>.<d010>.<d010>. <d050>Stop asking and it'll be fine.", 36, 1, true, { 255,255,255 });
+                arr->addObject(Object);   
+
+                break;
+
+            case 5:
+                Object = AttemptAtReversingDialogObject::create("The Shopkeeper", "I<d010>.<d010>.<d010>. I said stop!", 36, 1, true, { 255,255,255 });
+                arr->addObject(Object);   
+
+                break;
+
+            case 6:
+                Object = AttemptAtReversingDialogObject::create("The Shopkeeper", "It's best if you stop<d010>.<d010>.<d010>.", 35, 1, true, { 255,255,255 });
+                arr->addObject(Object);   
+
+                break;
+
+            case 7:
+                Object = AttemptAtReversingDialogObject::create("The Shopkeeper", ".<d010>.<d010>.", 34, 1, true, { 255,255,255 });
+                arr->addObject(Object);   
+
+                break;
+
+            case 8:
+                Object = AttemptAtReversingDialogObject::create("The Shopkeeper", "Oh no<d010>.<d010>.<d010>. not again.", 35, 1, true, { 255,255,255 });
+                arr->addObject(Object);
+
+                break;
+
+            case 9:
+                Object = AttemptAtReversingDialogObject::create("Zulguroth", "You DARE<d050> disturb a GOD!", 33, 1, true, { 255,255,255 });
+                arr->addObject(Object);
+
+                ss << "I have big plans for you, <d020><cl>";
+                ss << GameManager::get()->m_playerName.c_str();
+                ss << "</c>";
+
+                Object = AttemptAtReversingDialogObject::create("Zulguroth", ss.str(), 33, 1, true, { 255,255,255 });
+                arr->addObject(Object);
+
+                Object = AttemptAtReversingDialogObject::create("Zulguroth", "Now<d010>.<d010>.<d010>. BEGONE!", 33, 1, true, { 255,255,255 });
+                arr->addObject(Object);
+
+                break;
+
+            default:
+                break;
+            }
+
+            auto dl = DialogLayer::createDialogLayer(nullptr, arr, 2);
+            dl->animateIn(DialogAnimationType::FromLeft);
+            dl->setZOrder(999999);
+            CCScene::get()->addChild(dl);
+
+            secret++;
+
+            if (secret > 10)
+                secret = 0;
 
             return;
         }