+++ src/Replay/Replay.cpp
@@ -1,4 +1,44 @@
-/*#include "Macro.h"
+#include "Macro.h"
+
+using namespace gdr;
+
+struct MyInput : Input {
+	MyInput() = default;
+
+	MyInput(int frame, int button, bool player2, bool down, float xpos)
+		: Input(frame, button, player2, down) {}
+};
+
+struct MyReplay : Replay<MyReplay, MyInput> {
+	int attempts = -1;
+
+	MyReplay() : Replay("QOLMod Bot", "1.0") {}
+};
+
+/*class $modify (PlayerObject)
+{
+    virtual void update(float p0)
+    {
+        PlayerObject::update(p0);
+
+        //log::info("i hope this is tick instead of cocos update");
+
+        /*MyReplay r;
+
+		r.author = "camila314";
+		r.description = "we testing up in here";
+		r.attempts = 50;
+
+		r.inputs.push_back(MyInput(100, 1, false, true, 30.23));
+		r.inputs.push_back(MyInput(130, 1, false, false, 100.35));
+
+		std::vector<uint8_t> output = r.exportData(true);
+
+		std::string str(output.begin(), output.end());
+
+		log::info("replay data: {}", str);
+    }
+};*/
 
 class $modify (PauseLayer)
 {
@@ -6,9 +46,13 @@ class $modify (PauseLayer)
     {
         PauseLayer::customSetup();
 
-        //auto m = CCMenu::create(CCMenuItemSpriteExtra::create(CCSprite::createWithSpriteFrameName("GJ_achImage_001.png"), this, nullptr));
+		auto s = CCSprite::create("replay.png"_spr);
+		s->setScale(CCDirector::get()->getContentScaleFactor() * 0.5f);
 
-        //this->addChild(m);
-    }
-    //TodoReturn setupReplay(gd::string)
-};*/
+		auto btn = CCMenuItemSpriteExtra::create(s, this, nullptr);
+
+        auto m = CCMenu::create();
+		m->addChild(btn);
+        this->addChild(m);
+	}
+};